requirements.txt
RPi.GPIO
Adafruit_DHT
tensorflow
opencv-python
matplotlib
pandas
numpy
requests